requirement.dev.txt
numpy==1.21.5
pandas==1.3.5
psycopg==3.0.8
sshtunnel==0.4.0
PyYAML==6.0
pytz==2021.3
python_dateutil==2.8.2